Fix tests on PHP 7.0
