+++ src/Factory/DeviceDetectorFactory.php
@@ -2,7 +2,7 @@
 
 namespace Acsiomatic\DeviceDetectorBundle\Factory;
 
-use Acsiomatic\DeviceDetectorBundle\Bridge\DeviceDetector\LazyParserDeviceDetector;
+use Acsiomatic\DeviceDetectorBundle\Bridge\DeviceDetector\AutoParserDeviceDetector;
 use DeviceDetector\Cache\PSR6Bridge;
 use DeviceDetector\DeviceDetector;
 use Psr\Cache\CacheItemPoolInterface;
@@ -19,7 +19,7 @@ abstract class DeviceDetectorFactory
         bool $discardBotInformation,
         CacheItemPoolInterface $cache = null
     ): DeviceDetector {
-        $detector = new LazyParserDeviceDetector();
+        $detector = new AutoParserDeviceDetector();
 
         $detector->skipBotDetection($skipBotDetection);
         $detector->discardBotInformation($discardBotInformation);
